Added support for Arch Linux and rejigged the Unknown distro option to first, check for apt-get incase its a Debian variant but then failing if apt-get cannot be found.

+++ setup/install.sh
@@ -24,6 +24,11 @@ elif lsb_release -d | grep -q "Kali"; then
 	pip install pydispatcher
 	pip install flask
 	pip install pyOpenSSL
+elif lsb_release -d | grep -q "Arch"; then
+	Release=Arch
+	pacman -S --noconfirm --needed base-devel python2 python2-m2crypto swig python2-pip openssl python2-flask python2-pyopenssl python2-crypto
+	pip2 install iptools
+	pip2 install pydispatcher
 elif lsb_release -d | grep -q "Ubuntu"; then
 	Release=Ubuntu
 	apt-get install -y python-dev python-m2crypto swig python-pip libssl-dev
@@ -32,18 +37,25 @@ elif lsb_release -d | grep -q "Ubuntu"; then
 	pip install pydispatcher
 	pip install flask
 	pip install pyOpenSSL
+elif [ -f /usr/bin/apt-get ]; then
+	echo "Looks like a Debian/Ubuntu variant. Trying apt-get..."
+ 	apt-get install -y python-dev python-m2crypto swig python-pip libssl-dev
+ 	pip install pycrypto
+ 	pip install iptools
+ 	pip install pydispatcher
+ 	pip install flask
+ 	pip install pyOpenSSL
 else
-	echo "Unknown distro - Debian/Ubuntu Fallback"
-	 apt-get install -y python-dev python-m2crypto swig python-pip libssl-dev
-	 pip install pycrypto
-	 pip install iptools
-	 pip install pydispatcher
-	 pip install flask
-	 pip install pyOpenSSL
+	echo "Unknown distro - install cannot proceed..."
+	exit 1;
 fi
 
 # set up the database schema
-./setup_database.py
+if $Release == "Arch"; then
+	python2 ./setup_database.py
+else
+	./setup_database.py
+fi
 
 # generate a cert
 ./cert.sh
